Autoriser les actions avec les gates

+++ app/Http/Controllers/Admin/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\User;
 use App\Role;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class UsersController extends Controller
 {
@@ -67,6 +68,11 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        if (Gate::denies('edit-users')){
+
+            return redirect()->route('admin.users.index');
+        }
+         
         $roles = Role::all();
         return view('admin.users.edit',[
                 'user'=> $user,
@@ -96,6 +102,10 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+        if (Gate::denies('delete-users')){
+
+            return redirect()->route('admin.users.index');
+        }
         $user->roles()->detach();
         $user->delete();
